demo_generator/resources/requirements.txt
PyByteBuffer
jsons
web3
pandas
numpy
paho-mqtt==1.4.0